backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.1
bitarray==3.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cytoolz==1.0.0
dnspython==2.7.0
email-validator==2.2.0
eth-abi==5.1.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
fastapi==0.115.5
fastapi-cli==0.0.5
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
hexbytes==1.2.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
idna==3.10
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
parsimonious==0.10.0
passlib==1.7.4
prometheus-client==0.8.0
prometheus-fastapi-instrumentator==1.1.1
propcache==0.2.0
pycryptodome==3.21.0
pydantic==2.10.2
pydantic-core==2.27.1
pygments==2.18.0
pyjwt==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.18
pyunormalize==16.0.0
pywin32==308
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rlp==4.0.1
shellingham==1.5.4
sniffio==1.3.1
sqlalchemy==2.0.36
starlette==0.41.3
toolz==1.0.0
typer==0.14.0
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
watchfiles==1.0.0
web3==7.6.0
websockets==13.1
yarl==1.18.0
